% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_latex.R
\name{setup_latex}
\alias{setup_latex}
\title{Helper function to install latex if it isn't already installed}
\usage{
setup_latex(
  ...,
  packages = c("standalone", "pgf", "preview", "luatex85", "fontspec", "luaotfload",
    "latex-bin", "lm", "graphics", "grfext", "plex", "plex-otf", "unicode-math",
    "amsmath", "lualatex-math", "mathtools")
)
}
\arguments{
\item{...}{parameters that are passed on to \code{tinytex::install_tinytex(...)}}

\item{packages}{a character vector of Latex packages that are installed.
By default, a number of essential packages.}
}
\value{
\code{invisible(TRUE)} if the function runs through.
}
\description{
Helper function to install latex if it isn't already installed
}
