% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactplot.R
\name{theme_xp}
\alias{theme_xp}
\title{A simple theme for publications with few superfluous lines}
\usage{
theme_xp(
  fontsize = xp$fontsize,
  fontsize_small = xp$fontsize_small,
  fontsize_tiny = xp$fontsize_tiny,
  fontsize_large = xp$fontsize_large,
  line_size = 0.3,
  panel.spacing = unit(2, "mm")
)
}
\arguments{
\item{fontsize, fontsize_small, fontsize_tiny, fontsize_large}{the font sizes}
}
\description{
This function extends \code{\link[cowplot:theme_cowplot]{cowplot::theme_cowplot}}
}
