% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactplot.R
\docType{data}
\name{xp}
\alias{xp}
\title{Object that holds exactplot defaults}
\format{
An object of class \code{environment} of length 4.
}
\usage{
xp
}
\description{
See for example \code{xp$fontsize}
}
\keyword{datasets}
