% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactplot.R
\name{xp_add_latex_package}
\alias{xp_add_latex_package}
\title{Modify the tikzLualatexPackages option}
\usage{
xp_add_latex_package(packages)
}
\arguments{
\item{packages}{the name of the packages. Can be the full import statement (\verb{\\usePackage\{abcd\}})
in which case it is not modified or just the package name in which case it is wrapped in
\verb{\\usePackage\{...\}}.}
}
\description{
Modify the tikzLualatexPackages option
}
