% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_composition.R
\name{xp_compose_plots}
\alias{xp_compose_plots}
\alias{xp_plot}
\alias{xp_text}
\alias{xp_origin}
\alias{xp_graphic}
\title{Compose plots with millimeter precision}
\usage{
xp_compose_plots(
  ...,
  .plot_objs = NULL,
  width = 180,
  height = 110,
  units = c("mm", "cm", "inches", "px"),
  show_grid_lines = FALSE,
  keep_tex_file = FALSE,
  latex_engine = "luatex",
  filename = NULL
)

xp_plot(plot, x = 0, y = 0, width = NULL, height = NULL, units = NULL)

xp_text(label, x = 0, y = 0, fontsize = xp$fontsize, hjust = 0, vjust = 1, ...)

xp_origin(..., .plot_objs = NULL, x = 0, y = 0)

xp_graphic(
  filename,
  x = 0,
  y = 0,
  width = NULL,
  height = NULL,
  units = c("mm", "cm", "inches", "px"),
  anchor = c("north west", "south west", "base")
)
}
\arguments{
\item{...}{the panels (have to be grid objects)}

\item{.plot_objs}{alternative specification of the panels as a list}

\item{width, height}{the size of the total plot area. Defaults to the width of a plot in
Nature.}

\item{units}{the unit of the plot dimensions.}

\item{show_grid_lines}{boolean that indicates if helper lines are displayed in the background
to help align panels.}

\item{keep_tex_file}{boolean that indicates if the .tex file are retained after the latex
compilation completes. Only applies if \code{filename} is not \code{NULL}.}

\item{latex_engine}{the Latex engine used to compile the .tex document.}

\item{filename}{optional filename where the plot is saved. If \code{NULL} it is plotted without rendering
the latex.}

\item{x, y}{the x and y position relative to the top left corner.}

\item{fontsize}{the font size of the text.}
}
\description{
Compose plots with millimeter precision
}
\details{
Text that contains Latex code is only rendered when saving as a file.
}
