% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactplot.R
\name{xp_get_legend}
\alias{xp_get_legend}
\title{Get legend}
\usage{
xp_get_legend(plot)
}
\arguments{
\item{plot}{the plot from which the legend is extracted.}
}
\description{
Wraps \code{cowplot::get_plot_component}
}
