% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactplot.R
\name{xp_umap_axis}
\alias{xp_umap_axis}
\title{Add a pretty UMAP (or tSNE) axis to a plot}
\usage{
xp_umap_axis(
  label = "UMAP",
  fontsize = xp$fontsize_small,
  arrow_length = 10,
  label_offset = 1,
  fix_coord = TRUE,
  remove_axes = TRUE,
  arrow_spec = grid::arrow(ends = "both", type = "closed", angle = 20, length =
    unit(arrow_length/7, units)),
  units = "mm",
  ...
)
}
\arguments{
\item{label}{the axis label}

\item{fontsize}{the font size of the axis label}

\item{arrow_length}{length of the axis lines (in \code{units})}

\item{label_offset}{how far away from the axis the text is positioned (in \code{units})}

\item{fix_coord}{flag if a the aspect ratio is fixed to 1.}

\item{remove_axes}{if the original axes are removed by modifying the theme.}

\item{arrow_spec}{the definition of the \code{grid::arrow}}

\item{units}{the unit used for the arguments.}

\item{...}{additional arguments passed to \code{coord_fixed}.}
}
\description{
Add a pretty UMAP (or tSNE) axis to a plot
}
